requirements.txt
asgiref==3.2.10
colorama==0.4.4
Django==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
loguru==0.5.3
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
win32-setctime==1.0.3
